requirements.txt
aiogram==3.19.0
openai==1.82.0
python-dotenv==1.1.0
aiohttp==3.11.18
